requirements.txt
asgiref==3.8.1
autoflake==2.3.1
autopep8==2.2.0
black==24.4.2
Brotli==1.1.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
DateTime==5.5
distlib==0.3.8
dj-database-url==2.2.0
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
docformatter==1.7.5
filelock==3.15.3
flake8==7.0.0
gunicorn==22.0.0
h11==0.14.0
identify==2.5.36
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.0
pathspec==0.12.1
pip-check-reqs==2.5.3
platformdirs==4.2.2
pre-commit==3.7.1
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
setuptools==69.2.0
sqlparse==0.4.4
typing_extensions==4.12.2
tzdata==2024.1
untokenize==0.1.1
uvicorn==0.30.1
virtualenv==20.26.2
whitenoise==6.7.0
zope.interface==6.2
